xversion/zerochecksum/test_framework: Simplify version testing
Replace copies of custom version handler with new, more general
superclass.
